fully working parsing
