+++ src/components/flow/management/DeleteEdge.tsx
@@ -1,6 +1,5 @@
 import { BaseEdge, EdgeLabelRenderer, EdgeProps, getBezierPath, useReactFlow } from '@xyflow/react';
-import { FiX } from 'react-icons/fi';
-import { motion } from 'framer-motion';
+import { XIcon } from 'lucide-react';
 
 export default function FlowEdgeDeleteButton({
 	id,
@@ -35,36 +34,30 @@ export default function FlowEdgeDeleteButton({
 				path={edgePath}
 				markerEnd={markerEnd}
 				style={{
-					stroke: selected ? 'hsl(var(--primary))' : 'hsl(var(--muted-foreground))',
-					strokeWidth: selected ? 3 : 2,
-					transition: 'all 0.3s ease',
+					stroke: selected ? '#6e6a95' : '#908dae',
 					...style
 				}}
 			/>
-
 			<EdgeLabelRenderer>
-				<motion.button
-					initial={{ scale: 0 }}
-					animate={{ scale: 1 }}
-					whileHover={{ scale: 1.1 }}
-					whileTap={{ scale: 0.9 }}
+				<div
 					style={{
 						position: 'absolute',
-						left: `${labelX}px`,
-						top: `${labelY}px`,
-						transform: 'translate(-50%, -50%)',
+						transform: `translate(-50%, -50%) translate(${labelX}px,${labelY}px)`,
+						// everything inside EdgeLabelRenderer has no pointer events by default
+						// if you have an interactive element, set pointer-events: all
 						pointerEvents: 'all',
-						zIndex: 10
+						zIndex: 100
 					}}
-					className="nodrag nopan h-7 w-7 rounded-full bg-muted hover:bg-warning/80 border-2 border-warning/60 
-							 flex items-center justify-center shadow-lg transition-colors focus:outline-warning"
+					className="nodrag nopan cursor-pointer h-4 w-4 rounded-full flex items-center justify-center bg-muted cursor-pointer hover:cursor-pointer hover:bg-muted/50"
 					onClick={onEdgeClick}
-					title="Delete connection"
-					aria-label="Delete connection"
 				>
-					<FiX className="text-warning w-4 h-4 drop-shadow" />
-				</motion.button>
+					<XIcon className="h-3 w-3 text-foreground" />
+				</div>
 			</EdgeLabelRenderer>
+
+			<circle r="4" fill="#ff0073">
+				<animateMotion dur="1s" repeatCount="indefinite" path={edgePath} />
+			</circle>
 		</>
 	);
 }
